Use Standard Go Project Layout
Fixes #2
