+++ cmd/efactura-cli/cmd/api_xml2pdf.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/printesoi/e-factura-go/efactura"
+	"github.com/printesoi/e-factura-go/pkg/efactura"
 	"github.com/spf13/cobra"
 )
