+++ pkg/etransport/client_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License
 
-package efactura
+package etransport
 
 import (
 	"context"
@@ -20,25 +20,28 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"testing"
 
+	"github.com/google/uuid"
+	"github.com/stretchr/testify/assert"
 	xoauth2 "golang.org/x/oauth2"
 
-	"github.com/printesoi/e-factura-go/oauth2"
+	"github.com/printesoi/e-factura-go/pkg/oauth2"
 )
 
 func getTestCIF() string {
-	return os.Getenv("EFACTURA_TEST_CIF")
+	return os.Getenv("ETRANSPORT_TEST_CIF")
 }
 
 // setupTestEnvOAuth2Config creates a OAuth2Config from the environment.
 // If skipIfEmptyEnv if set to false and the env variables
-// EFACTURA_TEST_CLIENT_ID, EFACTURA_TEST_CLIENT_SECRET,
-// EFACTURA_TEST_REDIRECT_URL are not set, this method returns an error.
+// ETRANSPORT_TEST_CLIENT_ID, ETRANSPORT_TEST_CLIENT_SECRET,
+// ETRANSPORT_TEST_REDIRECT_URL are not set, this method returns an error.
 // If skipIfEmptyEnv is set to true and the env vars
 // are not set, this method returns a nil config.
 func setupTestEnvOAuth2Config(skipIfEmptyEnv bool) (oauth2Cfg *oauth2.Config, err error) {
-	clientID := os.Getenv("EFACTURA_TEST_CLIENT_ID")
-	clientSecret := os.Getenv("EFACTURA_TEST_CLIENT_SECRET")
+	clientID := os.Getenv("ETRANSPORT_TEST_CLIENT_ID")
+	clientSecret := os.Getenv("ETRANSPORT_TEST_CLIENT_SECRET")
 	if clientID == "" || clientSecret == "" {
 		if skipIfEmptyEnv {
 			return
@@ -47,7 +50,7 @@ func setupTestEnvOAuth2Config(skipIfEmptyEnv bool) (oauth2Cfg *oauth2.Config, er
 		return
 	}
 
-	redirectURL := os.Getenv("EFACTURA_TEST_REDIRECT_URL")
+	redirectURL := os.Getenv("ETRANSPORT_TEST_REDIRECT_URL")
 	if redirectURL == "" {
 		err = errors.New("invalid redirect URL")
 		return
@@ -79,7 +82,7 @@ func setupRealClient(skipIfEmptyEnv bool, oauth2Cfg *oauth2.Config) (*Client, er
 		}
 	}
 
-	tokenJSON := os.Getenv("EFACTURA_TEST_INITIAL_TOKEN_JSON")
+	tokenJSON := os.Getenv("ETRANSPORT_TEST_INITIAL_TOKEN_JSON")
 	if tokenJSON == "" {
 		return nil, errors.New("Invalid initial token json")
 	}
@@ -91,11 +94,57 @@ func setupRealClient(skipIfEmptyEnv bool, oauth2Cfg *oauth2.Config) (*Client, er
 
 	onTokenChanged := func(ctx context.Context, token *xoauth2.Token) error {
 		tokenJSON, _ := json.Marshal(token)
-		fmt.Printf("[E-FACTURA] token changed: %s\n", string(tokenJSON))
+		fmt.Printf("[E-TRANSPORT] token changed: %s\n", string(tokenJSON))
 		return nil
 	}
 
 	ctx := context.Background()
 	client, err := NewSandboxClient(ctx, oauth2Cfg.TokenSourceWithChangedHandler(ctx, token, onTokenChanged))
 	return client, err
+}
+
+func TestGenerateAuthCodeURL(t *testing.T) {
+	assert := assert.New(t)
+
+	cfg, err := setupTestEnvOAuth2Config(true)
+	if !assert.NoError(err) {
+		return
+	}
+	if cfg == nil {
+		t.Skipf("Skipping test, no credentials to create oauth2 config")
+	}
+
+	uuid, err := uuid.NewRandom()
+	if !assert.NoError(err) {
+		return
+	}
+
+	url := cfg.AuthCodeURL(uuid.String())
+	assert.True(url != "")
+	fmt.Printf("%s\n", url)
+}
+
+func TestExchangeCode(t *testing.T) {
+	assert := assert.New(t)
+
+	cfg, err := setupTestEnvOAuth2Config(true)
+	if !assert.NoError(err) {
+		return
+	}
+	if cfg == nil {
+		t.Skipf("Skipping test, no credentials to create oauth2 config")
+	}
+
+	code := os.Getenv("ETRANSPORT_TEST_EXCHANGE_CODE")
+	if code == "" {
+		t.Skipf("Skipping test, no exchange code")
+	}
+
+	token, err := cfg.Exchange(context.Background(), code)
+	if !assert.NoError(err) {
+		return
+	}
+
+	tj, _ := json.Marshal(token)
+	fmt.Printf("%s\n", string(tj))
 }
