+++ pkg/etransport/rest.go
@@ -23,8 +23,8 @@ import (
 	"strings"
 
 	ierrors "github.com/printesoi/e-factura-go/internal/errors"
-	"github.com/printesoi/e-factura-go/types"
-	ixml "github.com/printesoi/e-factura-go/xml"
+	"github.com/printesoi/e-factura-go/pkg/types"
+	ixml "github.com/printesoi/e-factura-go/pkg/xml"
 )
 
 const (
