+++ pkg/text/text.go
